Emory Dunn Var02 & Colin Reid Sound Var
For the second Variation of Week 02-03 I changed the shape of the Circle in the middle to a square shape and added a flow field. The lines start and end in a defined position, but their path changes and curves from the flow field, never knowing how lonng their adventure will be.

For Week 04 I decided to recreate Colin Reids cirlce collection and added the randomness of the colors, different spawning times so you can see when each circle is created and sounds to every one of them. Depending on their size, the sound is higher or lower.